Move to json
CSVs are great and all, but we need something more complex for things
like epochs and events (that are category-dependent). So now we move to
JSON, where we can reproduce that complexity.

I also tried yaml, but it was such a fight to even write valid yaml...
json is nicer