Completely rewrite everything
So, this is now a Javascript project (for now, but I'm thinking of
Kotlin). Bokeh was nice for a proof of concept and to get this thing
running, but it was ultimately too limiting. I couldn't really influence
the tooltips like I wanted, with different info for different object
types. Having labels for eras wasn't easy, as they covered some other
objects. In summary, I had too less control over the final product.

Enter Javascript with fabric.js. I'm not sure yet whether I like JS, but
I have total control over the rendering, and already some things are
better than before (tooltips, for example, or that nice bar highlighting
where you are). Of course, I have to do more manually - it _is_ more low
level, after all. But now I can think about showing more stuff when
you're zoomed in, for example. Or highlighting on mouseover. Or
whatever.

All in all, it was a lot of work, and lots of new stuff. But I quite
like the result 😄.